LPS-61455 Package moving
